orchestrator/requirements.txt
# Core dependencies
google-adk>=1.17.0
google-cloud-aiplatform>=1.38.0
google-cloud-run>=0.9.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Web/HTTP
httpx>=0.25.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI
click>=8.1.0
rich>=13.7.0

# Utilities
pyyaml>=6.0.1
jinja2>=3.1.2
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
mypy>=1.7.0
flake8>=6.1.0
